remove trait and move into separate files
