+++ lib/segment/src/index/field_index/map_index/mod.rs
@@ -1,11 +1,12 @@
-use std::collections::{BTreeSet, HashMap};
+pub mod mutable_map_index;
+
 use std::fmt::Display;
 use std::hash::Hash;
-use std::iter;
 use std::str::FromStr;
 use std::sync::Arc;
 
 use itertools::Itertools;
+use mutable_map_index::MutableMapIndex;
 use parking_lot::RwLock;
 use rocksdb::DB;
 use serde_json::Value;
@@ -26,209 +27,48 @@ use crate::types::{
 };
 use crate::vector_storage::div_ceil;
 
-/// HashMap-based type of index
-pub struct MutableMapIndex<N: Hash + Eq + Clone + Display + FromStr> {
-    map: HashMap<N, BTreeSet<PointOffsetType>>,
-    point_to_values: Vec<Vec<N>>,
-    /// Amount of point which have at least one indexed payload value
-    indexed_points: usize,
-    values_count: usize,
-    db_wrapper: DatabaseColumnWrapper,
-}
-
-pub enum MapIndexEnum<N: Hash + Eq + Clone + Display + FromStr> {
+pub enum MapIndex<N: Hash + Eq + Clone + Display + FromStr> {
     Mutable(MutableMapIndex<N>),
 }
 
-impl<N: Hash + Eq + Clone + Display + FromStr> MutableMapIndex<N> {
+impl<N: Hash + Eq + Clone + Display + FromStr> MapIndex<N> {
     pub fn new(db: Arc<RwLock<DB>>, field_name: &str) -> Self {
-        let store_cf_name = Self::storage_cf_name(field_name);
-        let db_wrapper = DatabaseColumnWrapper::new(db, &store_cf_name);
-        Self {
-            map: Default::default(),
-            point_to_values: Vec::new(),
-            indexed_points: 0,
-            values_count: 0,
-            db_wrapper,
-        }
+        MapIndex::Mutable(MutableMapIndex::new(db, field_name))
     }
 
-    fn add_many_to_map<Q>(&mut self, idx: PointOffsetType, values: Vec<Q>) -> OperationResult<()>
-    where
-        Q: Into<N>,
-    {
-        if values.is_empty() {
-            return Ok(());
-        }
-
-        self.values_count += values.len();
-        if self.point_to_values.len() <= idx as usize {
-            self.point_to_values.resize(idx as usize + 1, Vec::new())
-        }
-
-        self.point_to_values[idx as usize] = Vec::with_capacity(values.len());
-        for value in values {
-            let entry = self.map.entry(value.into());
-            self.point_to_values[idx as usize].push(entry.key().clone());
-            let db_record = Self::encode_db_record(entry.key(), idx);
-            entry.or_default().insert(idx);
-            self.db_wrapper.put(db_record, [])?;
-        }
-        self.indexed_points += 1;
-        Ok(())
-    }
-
-    fn remove_point(&mut self, idx: PointOffsetType) -> OperationResult<()> {
-        if self.point_to_values.len() <= idx as usize {
-            return Ok(());
-        }
-
-        let removed_values = std::mem::take(&mut self.point_to_values[idx as usize]);
-
-        if !removed_values.is_empty() {
-            self.indexed_points -= 1;
-        }
-        self.values_count -= removed_values.len();
-
-        for value in &removed_values {
-            if let Some(vals) = self.map.get_mut(value) {
-                vals.remove(&idx);
-            }
-            let key = Self::encode_db_record(value, idx);
-            self.db_wrapper.remove(key)?;
-        }
-
-        Ok(())
-    }
-}
-
-impl<N: Hash + Eq + Clone + Display + FromStr> MapIndex<N> for MutableMapIndex<N> {
-    fn get_db_wrapper(&self) -> &DatabaseColumnWrapper {
-        &self.db_wrapper
-    }
-
-    fn load_from_db(&mut self) -> OperationResult<bool> {
-        if !self.db_wrapper.has_column_family()? {
-            return Ok(false);
-        }
-        self.indexed_points = 0;
-        for (record, _) in self.db_wrapper.lock_db().iter()? {
-            let record = std::str::from_utf8(&record).map_err(|_| {
-                OperationError::service_error("Index load error: UTF8 error while DB parsing")
-            })?;
-            let (value, idx) = Self::decode_db_record(record)?;
-            if self.point_to_values.len() <= idx as usize {
-                self.point_to_values.resize(idx as usize + 1, Vec::new())
-            }
-            if self.point_to_values[idx as usize].is_empty() {
-                self.indexed_points += 1;
-            }
-            self.values_count += 1;
-
-            let entry = self.map.entry(value);
-            self.point_to_values[idx as usize].push(entry.key().clone());
-            entry.or_default().insert(idx);
-        }
-        Ok(true)
-    }
-
-    fn get_values(&self, idx: PointOffsetType) -> Option<&[N]> {
-        self.point_to_values.get(idx as usize).map(|v| v.as_slice())
-    }
-
-    fn get_indexed_points(&self) -> usize {
-        self.indexed_points
-    }
-
-    fn get_values_count(&self) -> usize {
-        self.values_count
-    }
-
-    fn get_unique_values_count(&self) -> usize {
-        self.map.len()
-    }
-
-    fn get_points_with_value_count<Q>(&self, value: &Q) -> Option<usize>
-    where
-        Q: ?Sized,
-        N: std::borrow::Borrow<Q>,
-        Q: Hash + Eq,
-    {
-        self.map.get(value).map(|p| p.len())
-    }
-
-    fn get_iterator<Q>(&self, value: &Q) -> Box<dyn Iterator<Item = PointOffsetType> + '_>
-    where
-        Q: ?Sized,
-        N: std::borrow::Borrow<Q>,
-        Q: Hash + Eq,
-    {
-        self.map
-            .get(value)
-            .map(|ids| Box::new(ids.iter().copied()) as Box<dyn Iterator<Item = PointOffsetType>>)
-            .unwrap_or_else(|| Box::new(iter::empty::<PointOffsetType>()))
-    }
-
-    fn get_values_iterator(&self) -> Box<dyn Iterator<Item = &N> + '_> {
-        Box::new(self.map.keys())
-    }
-
-    fn except_iterator<'a, Q>(
-        &'a self,
-        excluded: &'a [Q],
-    ) -> Box<dyn Iterator<Item = PointOffsetType> + 'a>
-    where
-        Q: PartialEq<N>,
-    {
-        Box::new(
-            self.get_values_iterator()
-                .filter(|key| !excluded.iter().any(|e| e.eq(*key)))
-                .flat_map(|key| self.get_iterator(key))
-                .unique(),
-        )
-    }
-}
-
-impl<N: Hash + Eq + Clone + Display + FromStr> MapIndexEnum<N> {
-    pub fn new(db: Arc<RwLock<DB>>, field_name: &str) -> Self {
-        MapIndexEnum::Mutable(MutableMapIndex::new(db, field_name))
-    }
-}
-
-impl<N: Hash + Eq + Clone + Display + FromStr> MapIndex<N> for MapIndexEnum<N> {
     fn get_db_wrapper(&self) -> &DatabaseColumnWrapper {
         match self {
-            MapIndexEnum::Mutable(index) => index.get_db_wrapper(),
+            MapIndex::Mutable(index) => index.get_db_wrapper(),
         }
     }
 
     fn load_from_db(&mut self) -> OperationResult<bool> {
         match self {
-            MapIndexEnum::Mutable(index) => index.load_from_db(),
+            MapIndex::Mutable(index) => index.load_from_db(),
         }
     }
 
-    fn get_values(&self, idx: PointOffsetType) -> Option<&[N]> {
+    pub fn get_values(&self, idx: PointOffsetType) -> Option<&[N]> {
         match self {
-            MapIndexEnum::Mutable(index) => index.get_values(idx),
+            MapIndex::Mutable(index) => index.get_values(idx),
         }
     }
 
     fn get_indexed_points(&self) -> usize {
         match self {
-            MapIndexEnum::Mutable(index) => index.get_indexed_points(),
+            MapIndex::Mutable(index) => index.get_indexed_points(),
         }
     }
 
     fn get_values_count(&self) -> usize {
         match self {
-            MapIndexEnum::Mutable(index) => index.get_values_count(),
+            MapIndex::Mutable(index) => index.get_values_count(),
         }
     }
 
     fn get_unique_values_count(&self) -> usize {
         match self {
-            MapIndexEnum::Mutable(index) => index.get_unique_values_count(),
+            MapIndex::Mutable(index) => index.get_unique_values_count(),
         }
     }
 
@@ -239,7 +79,7 @@ impl<N: Hash + Eq + Clone + Display + FromStr> MapIndex<N> for MapIndexEnum<N> {
         Q: Hash + Eq,
     {
         match self {
-            MapIndexEnum::Mutable(index) => index.get_points_with_value_count(value),
+            MapIndex::Mutable(index) => index.get_points_with_value_count(value),
         }
     }
 
@@ -250,41 +90,23 @@ impl<N: Hash + Eq + Clone + Display + FromStr> MapIndex<N> for MapIndexEnum<N> {
         Q: Hash + Eq,
     {
         match self {
-            MapIndexEnum::Mutable(index) => index.get_iterator(value),
+            MapIndex::Mutable(index) => index.get_iterator(value),
         }
     }
 
     fn get_values_iterator(&self) -> Box<dyn Iterator<Item = &N> + '_> {
         match self {
-            MapIndexEnum::Mutable(index) => index.get_values_iterator(),
+            MapIndex::Mutable(index) => index.get_values_iterator(),
         }
     }
 
-    fn except_iterator<'a, Q>(
-        &'a self,
-        excluded: &'a [Q],
-    ) -> Box<dyn Iterator<Item = PointOffsetType> + 'a>
-    where
-        Q: PartialEq<N>,
-    {
-        match self {
-            MapIndexEnum::Mutable(index) => index.except_iterator(excluded),
-        }
-    }
-}
-
-pub trait MapIndex<N: Hash + Eq + Clone + Display + FromStr> {
-    fn get_db_wrapper(&self) -> &DatabaseColumnWrapper;
-
-    fn storage_cf_name(field: &str) -> String {
+    pub fn storage_cf_name(field: &str) -> String {
         format!("{field}_map")
     }
 
-    fn recreate(&self) -> OperationResult<()> {
+    pub fn recreate(&self) -> OperationResult<()> {
         self.get_db_wrapper().recreate_column_family()
     }
-
-    fn load_from_db(&mut self) -> OperationResult<bool>;
 
     fn flusher(&self) -> Flusher {
         self.get_db_wrapper().flusher()
@@ -301,21 +123,7 @@ pub trait MapIndex<N: Hash + Eq + Clone + Display + FromStr> {
         CardinalityEstimation::exact(values_count)
     }
 
-    fn get_values(&self, idx: PointOffsetType) -> Option<&[N]>;
-
-    fn get_indexed_points(&self) -> usize;
-
-    fn get_values_count(&self) -> usize;
-
-    fn get_unique_values_count(&self) -> usize;
-
-    fn get_points_with_value_count<Q>(&self, value: &Q) -> Option<usize>
-    where
-        Q: ?Sized,
-        N: std::borrow::Borrow<Q>,
-        Q: Hash + Eq;
-
-    fn get_telemetry_data(&self) -> PayloadIndexTelemetry {
+    pub fn get_telemetry_data(&self) -> PayloadIndexTelemetry {
         PayloadIndexTelemetry {
             field_name: None,
             points_count: self.get_indexed_points(),
@@ -324,19 +132,11 @@ pub trait MapIndex<N: Hash + Eq + Clone + Display + FromStr> {
         }
     }
 
-    fn get_iterator<Q>(&self, value: &Q) -> Box<dyn Iterator<Item = PointOffsetType> + '_>
-    where
-        Q: ?Sized,
-        N: std::borrow::Borrow<Q>,
-        Q: Hash + Eq;
-
-    fn get_values_iterator(&self) -> Box<dyn Iterator<Item = &N> + '_>;
-
-    fn encode_db_record(value: &N, idx: PointOffsetType) -> String {
+    pub fn encode_db_record(value: &N, idx: PointOffsetType) -> String {
         format!("{value}/{idx}")
     }
 
-    fn decode_db_record(s: &str) -> OperationResult<(N, PointOffsetType)> {
+    pub fn decode_db_record(s: &str) -> OperationResult<(N, PointOffsetType)> {
         const DECODE_ERR: &str = "Index db parsing error: wrong data format";
         let separator_pos = s
             .rfind('/')
@@ -353,11 +153,11 @@ pub trait MapIndex<N: Hash + Eq + Clone + Display + FromStr> {
         Ok((value, idx))
     }
 
-    fn values_count(&self, point_id: PointOffsetType) -> usize {
+    pub fn values_count(&self, point_id: PointOffsetType) -> usize {
         self.get_values(point_id).map(|x| x.len()).unwrap_or(0)
     }
 
-    fn values_is_empty(&self, point_id: PointOffsetType) -> bool {
+    pub fn values_is_empty(&self, point_id: PointOffsetType) -> bool {
         self.get_values(point_id)
             .map(|x| x.is_empty())
             .unwrap_or(true)
@@ -480,10 +280,18 @@ pub trait MapIndex<N: Hash + Eq + Clone + Display + FromStr> {
         excluded: &'a [Q],
     ) -> Box<dyn Iterator<Item = PointOffsetType> + 'a>
     where
-        Q: PartialEq<N>;
+        Q: PartialEq<N>,
+    {
+        Box::new(
+            self.get_values_iterator()
+                .filter(|key| !excluded.iter().any(|e| e.eq(*key)))
+                .flat_map(|key| self.get_iterator(key))
+                .unique(),
+        )
+    }
 }
 
-impl PayloadFieldIndex for MapIndexEnum<SmolStr> {
+impl PayloadFieldIndex for MapIndex<SmolStr> {
     fn count_indexed_points(&self) -> usize {
         self.get_indexed_points()
     }
@@ -570,7 +378,7 @@ impl PayloadFieldIndex for MapIndexEnum<SmolStr> {
     }
 }
 
-impl PayloadFieldIndex for MapIndexEnum<IntPayloadType> {
+impl PayloadFieldIndex for MapIndex<IntPayloadType> {
     fn count_indexed_points(&self) -> usize {
         self.get_indexed_points()
     }
@@ -659,10 +467,10 @@ impl PayloadFieldIndex for MapIndexEnum<IntPayloadType> {
     }
 }
 
-impl ValueIndexer<String> for MapIndexEnum<SmolStr> {
+impl ValueIndexer<String> for MapIndex<SmolStr> {
     fn add_many(&mut self, id: PointOffsetType, values: Vec<String>) -> OperationResult<()> {
         match self {
-            MapIndexEnum::Mutable(index) => index.add_many_to_map(id, values),
+            MapIndex::Mutable(index) => index.add_many_to_map(id, values),
         }
     }
 
@@ -675,19 +483,19 @@ impl ValueIndexer<String> for MapIndexEnum<SmolStr> {
 
     fn remove_point(&mut self, id: PointOffsetType) -> OperationResult<()> {
         match self {
-            MapIndexEnum::Mutable(index) => index.remove_point(id),
+            MapIndex::Mutable(index) => index.remove_point(id),
         }
     }
 }
 
-impl ValueIndexer<IntPayloadType> for MapIndexEnum<IntPayloadType> {
+impl ValueIndexer<IntPayloadType> for MapIndex<IntPayloadType> {
     fn add_many(
         &mut self,
         id: PointOffsetType,
         values: Vec<IntPayloadType>,
     ) -> OperationResult<()> {
         match self {
-            MapIndexEnum::Mutable(index) => index.add_many_to_map(id, values),
+            MapIndex::Mutable(index) => index.add_many_to_map(id, values),
         }
     }
 
@@ -700,7 +508,7 @@ impl ValueIndexer<IntPayloadType> for MapIndexEnum<IntPayloadType> {
 
     fn remove_point(&mut self, id: PointOffsetType) -> OperationResult<()> {
         match self {
-            MapIndexEnum::Mutable(index) => index.remove_point(id),
+            MapIndex::Mutable(index) => index.remove_point(id),
         }
     }
 }
@@ -723,11 +531,11 @@ mod tests {
         data: &[Vec<N>],
         path: &Path,
     ) {
-        let mut index = MapIndexEnum::<N>::new(open_db_with_existing_cf(path).unwrap(), FIELD_NAME);
+        let mut index = MapIndex::<N>::new(open_db_with_existing_cf(path).unwrap(), FIELD_NAME);
         index.recreate().unwrap();
         for (idx, values) in data.iter().enumerate() {
             match &mut index {
-                MapIndexEnum::Mutable(index) => index
+                MapIndex::Mutable(index) => index
                     .add_many_to_map(idx as PointOffsetType, values.clone())
                     .unwrap(),
             }
@@ -739,7 +547,7 @@ mod tests {
         data: &[Vec<N>],
         path: &Path,
     ) {
-        let mut index = MapIndexEnum::<N>::new(open_db_with_existing_cf(path).unwrap(), FIELD_NAME);
+        let mut index = MapIndex::<N>::new(open_db_with_existing_cf(path).unwrap(), FIELD_NAME);
         index.load_from_db().unwrap();
         for (idx, values) in data.iter().enumerate() {
             let index_values: HashSet<N> = HashSet::from_iter(
