+++ lib/segment/src/index/query_optimization/condition_converter.rs
@@ -4,7 +4,6 @@ use serde_json::Value;
 
 use crate::common::utils::IndexesMap;
 use crate::id_tracker::IdTrackerSS;
-use crate::index::field_index::map_index::MapIndex;
 use crate::index::field_index::FieldIndex;
 use crate::index::query_optimization::optimized_filter::ConditionCheckerFn;
 use crate::index::query_optimization::payload_provider::PayloadProvider;
